requirements.txt
torch>=2.1.0
numpy>=1.24,<3.0
stable-baselines3>=2.3.0,<3.0.0
gymnasium[atari]>=0.29
shimmy>=2.0.0
autorom[accept-rom-license]>=0.6.1
ale-py>=0.8.1
tensorboard>=2.14.0
matplotlib>=3.7.0
tqdm>=4.66.0